Added Test and added logic in user and loan, Tested getting loan in postman and workd

+++ src/main/java/com/example/Gruppuppgift_Statsbibloteket/model/Users.java
@@ -7,15 +7,17 @@ import lombok.Setter;
 @Entity
 @Setter
 @Getter
-@Table(name="users")
+@Table(name = "users")
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int user_id;
+    private Long user_id;  // Changed Integer to Long
+
     private String first_name;
     private String last_name;
     private String email;
     private String member_number;
+
     public Users() {}
 
     public Users(String first_name, String last_name, String email, String member_number) {
